requirements.txt
certifi==2018.11.29
chardet==3.0.4
defusedxml==0.5.0
dj-database-url==0.5.0
Django==2.1.3
django-allauth==0.38.0
django-crispy-forms==1.7.2
django-js-asset==1.1.0
django-mptt==0.9.1
django-summernote==0.8.10.0
idna==2.7
oauthlib==2.1.0
Pillow==5.3.0
psycopg2==2.7.7
python-decouple==3.1
python3-openid==3.1.0
pytz==2018.7
requests==2.20.1
requests-oauthlib==1.0.0
urllib3==1.24.1
